requirements.txt
requests
openai
fastapi
uvicorn[standard]
python-dotenv
httpx
sseclient-py
pytest
pytest-asyncio
